requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.30.0
beautifulsoup4==4.12.2
lxml==4.9.2
celery==5.3.0
redis==4.5.5
azure-cosmos==4.3.1
azure-storage-blob==12.16.0
